requirements.txt
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
imageio==2.22.1
imutils==0.5.4
kiwisolver==1.4.4
matplotlib==3.6.0
networkx==2.8.7
numpy==1.23.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
scikit-image==0.19.3
scipy==1.9.1
Shapely==1.8.5.post1
six==1.16.0
tifffile==2022.10.10
tqdm==4.64.1
utils==1.0.1
